fix: keep file loading out of input factory
